Redirect documentation link to deepwiki
The deep wiki is so much more thorough. I was confused why I went to a barebones github pages instead of that when clicking on the first link that would be evident to me as comprehensive.

The github pages link remains as #10 if needed to be referenced in the future.